move tests to `tests` dir
