+++ cot/Cargo.toml
@@ -25,7 +25,7 @@ blake3.workspace = true
 bytes.workspace = true
 chrono = { workspace = true, features = ["alloc", "serde", "clock"] }
 chrono-tz.workspace = true
-clap = {workspace = true, features = ["string"] }
+clap = { workspace = true, features = ["string"] }
 cot_core.workspace = true
 cot_macros.workspace = true
 deadpool-redis = { workspace = true, features = ["tokio-comp", "rt_tokio_1"], optional = true }
